refactor: Racing -> Car로 변경
